Moved pages to help layout.js
Moved pages into routes that do not affect the pages on the frontend. This was done to use a layout file unique to pages that are accessible after a user logs in.
