+++ frontend/src/app/(with-layout)/hubs/[id]/page.js
