+++ frontend/src/app/(with-layout)/layout.js
@@ -1,18 +1,18 @@
 /*CAME WITH THIS IN CREATE NEXT APP*/
 import localFont from "next/font/local";
-import "./globals.css";
+import "../globals.css";
 import RebelHubNavBar from "@/components/navbar/RebelHubNavBar";
 import Sidebar from "@/components/sidebar/sidebar";
 
 import margins from "@/utils/Margins.module.css";
 
 const geistSans = localFont({
-  src: "./fonts/GeistVF.woff",
+  src: "../fonts/GeistVF.woff",
   variable: "--font-geist-sans",
   weight: "100 900",
 });
 const geistMono = localFont({
-  src: "./fonts/GeistMonoVF.woff",
+  src: "../fonts/GeistMonoVF.woff",
   variable: "--font-geist-mono",
   weight: "100 900",
 });
@@ -25,7 +25,7 @@ export const metadata = {
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
-      <body className={`${geistSans.variable} ${geistMono.variable}`}>
+      <body style={{'backgroundColor': '#000000'}} className={`${geistSans.variable} ${geistMono.variable}`}>
         <RebelHubNavBar/>
         <Sidebar/>
         <div className={margins.margins}>
