+++ frontend/src/app/(with-layout)/posts/[id]/page.js
