+++ frontend/src/app/(with-layout)/posts/create/page.js
